requirements.txt
# Batch Processing Dependencies
# Install with: pip install -r requirements_batch.txt

# Existing dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-ai>=0.0.1
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# NEW: Batch processing dependencies
google-generativeai>=0.3.0  # Google Gemini Batch API

# Optional: Redis for advanced caching
redis>=5.0.0

# Database client for entity storage
supabase>=2.0.0
